exclude hostname verification disabled IT from jenkins
