+++ src/test/resources/ssl/instances.yml
@@ -2,7 +2,8 @@ instances:
   - name: elasticsearch
     dns:
       - elasticsearch
-      - localhost
+    ip:
+      - ipAddress
   - name: kibana
     dns:
       - kibana
